# swag/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API для сервиса mpit2026-reg.\n\nОсновные возможности:\n- авторизация через MAX mini app и локальный JWT\n- получение и обновление профиля текущего пользователя\n- создание событий, получение invite token и вход по нему\n- управление гостями и фиксация статуса посещения\n- wishlist и фото внутри события\n\nАутентификация:\n- защищённые методы требуют заголовок Authorization: Bearer \u003cjwt\u003e\n- JWT можно получить через /auth/max/login или через мобильный flow /auth/max/start -\u003e /auth/max/complete -\u003e /auth/max/exchange\n\nSwagger UI доступен по адресу /api/docs/.",
        "title": "mpit2026-reg API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "mpit-bot.kostya1024.ru",
    "basePath": "/api/v1",
    "paths": {
...
        "/events": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт новое событие от имени текущего авторизованного пользователя.\nОрганизатором события автоматически становится пользователь из Bearer JWT.\nВ запросе нужно передать базовые параметры события: город, дату, время, ожидаемое число гостей и бюджет.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Создать событие",
                "parameters": [
                    {
                        "description": "Тело запроса с параметрами события",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/action.CreateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Событие успешно создано",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/core.Event"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Некорректный JSON или ошибка валидации полей",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Конфликт при создании события",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/events/join-by-token": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавляет текущего пользователя в событие по invite token.\nТокен должен быть активным и принадлежать существующему событию.\nПосле успешного входа backend возвращает карточку события с ролью доступа пользователя в этом событии.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Войти в событие по invite token",
                "parameters": [
                    {
                        "description": "Тело запроса с invite token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/action.JoinEventByTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно добавлен в событие",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/core.Event"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Некорректный JSON или пустой token",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Invite token не найден или недействителен",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Конфликт при присоединении к событию",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/events/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список событий, к которым текущий пользователь имеет отношение.\nВ выборку попадают события, где пользователь является организатором, соорганизатором или гостем.\nДля каждого события backend дополнительно возвращает access_role и связанные статусы гостя, если они применимы.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Получить список моих событий",
                "responses": {
                    "200": {
                        "description": "Список событий текущего пользователя",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/core.Event"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/events/{eventID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает полную карточку события по его идентификатору.\nДоступ к методу ограничен middleware авторизации и проверкой прав доступа к конкретному событию.\nЕсли у пользователя нет права читать событие, backend вернёт 403.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Получить событие по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор события",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Карточка события",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/core.Event"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Некорректный eventID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав для чтения события",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Событие не найдено",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/events/{eventID}/guests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список гостей, связанных с указанным событием.\nМожно дополнительно отфильтровать список по approval_status через query-параметр.\nДоступ к методу ограничен проверкой прав на чтение события.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Получить список гостей события",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор события",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по статусу подтверждения: pending, approved, rejected",
                        "name": "approval_status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список гостей события",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/core.EventGuest"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав для просмотра гостей",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/events/{eventID}/guests/{guestID}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет attendance_status конкретного гостя в рамках события.\nМетод предназначен для сценария, когда сам пользователь подтверждает или отклоняет своё участие.\nОрганизатор и co_host не могут использовать этот endpoint для управления посещаемостью других пользователей.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Обновить статус посещения гостя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор события",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор гостя",
                        "name": "guestID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса с attendance_status",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/action.UpdateGuestStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус посещения успешно обновлён",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/core.EventGuest"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Некорректный JSON, eventID, guestID или attendance_status",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав для обновления статуса посещения",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Гость не найден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/events/{eventID}/invite": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает активный invite token для указанного события.\nОбычно используется организатором или соорганизатором, чтобы передать ссылку или token приглашённому пользователю.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Получить invite token события",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор события",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Активный invite token события",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав для просмотра invite token",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Событие или invite token не найдены",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "action.CreateEventRequest": {
            "type": "object",
            "properties": {
                "budget": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "event_date": {
                    "type": "string"
                },
                "event_time": {
                    "type": "string"
                },
                "expected_guest_count": {
                    "type": "integer"
                }
            }
        },
...
        "action.JoinEventByTokenRequest": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "action.UpdateGuestStatusRequest": {
            "type": "object",
            "properties": {
                "attendance_status": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorBody"
                }
            }
        },
        "response.SuccessEnvelope": {
            "type": "object",
            "properties": {
                "data": {}
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}